implemented backward compatible jobservice for android 8
In order to save battery power, Google forces apps more and more to
become inactive as long as a device is inactive. The AlarmManager is no
longer suited to check regularly for new articles. By implementing
JobService instead of AlarmManager this can be overcome. Unfortunately,
JobService is not backward compatible, so we had to do that ourselves.
With this version, for older devices the AlarmManager is used, where on
new devices the JobService is used.